+++ src/components/MultiSelectChipComponent.tsx
@@ -1,0 +1,59 @@
+import Box from "@mui/material/Box";
+import OutlinedInput from "@mui/material/OutlinedInput";
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import FormControl from "@mui/material/FormControl";
+import Select from "@mui/material/Select";
+import Chip from "@mui/material/Chip";
+
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
+
+export default function MultiSelectChipComponent({
+  selectedItems = [],
+  setSelectedItems,
+  items,
+  label,
+}) {
+  const handleChange = (event) => {
+    setSelectedItems("breeds", event.target.value);
+  };
+
+  return (
+    <div>
+      <FormControl sx={{ width: "100%" }} size="medium">
+        <InputLabel id="multiple-chip-label">Select {label}</InputLabel>
+        <Select
+          labelId="multiple-chip-label"
+          id="multiple-chip"
+          multiple
+          value={selectedItems}
+          onChange={handleChange}
+          input={<OutlinedInput id="select-multiple-chip" label="Chip" />}
+          renderValue={(selected) => (
+            <Box sx={{ display: "flex", flexWrap: "wrap", gap: 0.5 }}>
+              {selected.map((value) => (
+                <Chip key={value} label={value} />
+              ))}
+            </Box>
+          )}
+          MenuProps={MenuProps}
+        >
+          {items.map((item) => (
+            <MenuItem key={item} value={item}>
+              {item}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+    </div>
+  );
+}
